requirements.txt
deepface
opencv-python
transformers
torch
streamlit
Pillow
pandas
librosa
sounddevice
scipy